Renamed packages in order to change case
Added packages from "download" directory

+++ Scripts/build.chocolatey-package.extension.ps1
@@ -1,11 +1,13 @@
 $packagesDir = Join-Path $PSScriptRoot '..\..\..\BoxStarter' -Resolve
 $parentDir = Join-Path $PSScriptRoot '..\' -Resolve
 
-choco uninstall Chocolatey-Package.extension -f
-Remove-Item -Recurse C:\ProgramData\Chocolatey\extensions\chocolatey-Package -ErrorAction SilentlyContinue
+if ((choco list Chocolatey-Package -lo -r)) {
+    choco uninstall Chocolatey-Package.extension -f
+    Remove-Item -Recurse C:\ProgramData\Chocolatey\extensions\Chocolatey-Package -ErrorAction SilentlyContinue
+}
 
 & $parentDir\build-push.ps1 Chocolatey-Package.extension -f
 
-choco install Chocolatey-Package.extension -f -s $packagesDir
+choco upgrade Chocolatey-Package.extension -f -s $packagesDir
 
-Import-Module 'C:\ProgramData\Chocolatey\extensions\chocolatey-package\Chocolatey-Package.extension.psm1' -Force
+Import-Module 'C:\ProgramData\Chocolatey\extensions\Chocolatey-Package\Chocolatey-Package.extension.psm1' -Force
